% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sampling.R
\name{rejection_sampling}
\alias{rejection_sampling}
\title{Single Variable Rejection Sampling}
\usage{
rejection_sampling(n = 1, pdf, a = 0, b = 1, C)
}
\arguments{
\item{n}{is the number of samples to generate.}

\item{pdf}{is the pdf of the random variable you wish to sample from.}

\item{a}{is the lower bound of the random variable you wish to sample from.}

\item{b}{is the upper bound of the random variable you wish to sample from.}

\item{C}{is a numeric such that f(x) <= C for all values of x.}
}
\value{
my_sample A vector containg the desired number of random samples from the desired distribution.
}
\description{
This is the main function of the MATH 4800 project.
It recieves a function's pdf and generates a number of samples from the rv with the provided pdf using rejection sampling.
It is thereby somewhat similar to functions such as rbinom, runif, etc.
}
\examples{
rejection_sampling(10^4, function(x)1, 0, 1, 1)
rejection_sampling(10^4, function(x)2*x, 0, 1, 2)
rejection_sampling(10^4, function(x)exp(-x), 0, 10, 1)
}
\author{
Peter Reiker
}
